% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertXFrag2Mgf.R
\name{convertXFrag2Mgf}
\alias{convertXFrag2Mgf}
\title{Convert \code{xcmsFragments} object into a list (MGF format)}
\usage{
convertXFrag2Mgf(xfrag)
}
\arguments{
\item{xfrag}{\code{xcmsFragments} object}
}
\value{
Returns a nested list, each of which contains \code{precursor} (numeric)
     and \code{spec} (data.frame, with \code{mz} and \code{intensity} columns).
}
\description{
Convert \code{xcmsFragments} object into a list (MGF format)
}

