% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLXneg.R
\name{correctLXneg}
\alias{correctLXneg}
\title{Species ratio correction for LipidXplorer identification result (negative mode)}
\usage{
correctLXneg(df, anID = NULL, precID = NULL, faID = NULL)
}
\arguments{
\item{df}{A data frame. Result of LipidXplorer identification, without unnecessary lines.
E.g., MFQL title lines that starts with ### should be removed.}

\item{anID}{Numeric vector. Column indexes for annotation part of input data frame.}

\item{precID}{Numeric vector. Column indexes for precursor ion intensities.}

\item{faID}{Numeric vector. Column indexes for fatty acid anion intensities.}
}
\value{
A data frame with annotation part, corrected precursor intensities,
     and original fatty acid anion intensities.
}
\description{
This function performs intensity correction based on intensities of precursor ion
     intensity and fatty acid anions. In a situation where there are several lipid
     species of the same m/z value but with different fatty acid compositions, this
     function allocates precursor ion intensity based on fatty acid fragment ion
     intensities.
}
\references{
Surma MA, Herzog R, Vasilj A, Klose C, Christinat N,
     Morin Rivron D, Simons K, Masoodi M & Sampaio JL (2015) An automated shotgun
     lipidomics platform for high throughput, comprehensive, and quantitative
     analysis of blood plasma intact lipids. European Journal of Lipid Science and
     Technology 117: 1540–1549
}

