% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctPrecursorMz.R
\name{correctPrecursorMz}
\alias{correctPrecursorMz}
\title{Precursor m/z correction of xcmsRaw object imported from mzML (converted from Thermo RAW)}
\usage{
correctPrecursorMz(xraw, path, mzabs = 0.05)
}
\arguments{
\item{xraw}{An \code{xcmsRaw} object.}

\item{path}{A path to the original mzML file.}

\item{mzabs}{Window size to search real precursor m/z value (default 0.05).}
}
\value{
\code{xcmsRaw} object with corrected precursor m/z values
}
\description{
This function is a kind of a bug fix for mzML (generated from ThermoRAW by
   msConvert) data acquired on LTQ-Orbitrap XL, with monoisotopic selection
   OFF under DDA settings. The setting results in the precursor mz entry in
   the mzML to be "monoisotopic mass" (even though "monoisotopic selection"
   is disabled). Therefore real precursor mz values are not read from the
   mzML file into xcmsRaw object. As the real precursor m/z values are still
   present in the mzML file (though only to 2 decimals), this function extract
   the real precursor m/z and pick peak near the m/z value.
}

