% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findScans.R
\name{findScans}
\alias{findScans}
\title{A function that returns scan indexes from a precursor m/z value and xcmsRaw object}
\usage{
findScans(prmz, xraw, mzabs = 0.1)
}
\arguments{
\item{prmz}{A numeric value. Numeric vectors cannot be used
(in that cases, only the first element will be used).}

\item{xraw}{An object of class \code{xcmsRaw}, which is read into by \code{xcmsRaw()}
with \code{includeMSn = TRUE} option.}

\item{mzabs}{Search tolerance in m/z (not ppm).}
}
\value{
Returns a data frame, with columns \code{PrecursorMz}, \code{ScanNum},
     and \code{ScanNumMS2}.
}
\description{
This function retrieves scan indexes ("real scan number" and scan number within MS2 scans)
     from precursor ion list (from meta-data).
}

