% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeaksAvgMS1.R
\name{getPeaksAvgMS1}
\alias{getPeaksAvgMS1}
\title{A function to pick peaks from an averaged MS1 spectrum from a \code{xcmsRaw} object}
\usage{
getPeaksAvgMS1(xcmsraw, scans = NULL, mzrange = NULL, ylim = NULL,
  ad.lab.mzrange = NA, ad.top = 3, top = 12, halfWindowSize = NULL,
  SNR = 50, digits = 4, ...)
}
\arguments{
\item{xcmsraw}{\code{xcmsRaw} object.}

\item{scans}{A numeric vector of scan indexes. If provided, only the indicated scans are averaged and plotted.}

\item{mzrange}{m/z range to plot.}

\item{ylim}{Y axis range to plot.}

\item{ad.top}{The number of additional peak labels.}

\item{top}{The number of peak labels.}

\item{halfWindowSize}{See \code{MALDIquant::labelPeaks()} for detail.}

\item{SNR}{See \code{MALDIquant::labelPeaks()} for detail.}

\item{digits}{Digits for peak labels.}

\item{ad.lab.}{mzrange m/z range to add additional peak labels.}
}
\value{
Top N peaks as a data frame.
}
\description{
This function accepts \code{xcmsRaw} object with MS/MS data (direct infusion,
   data-dependent acquisition) and return top N peaks. Correspond to \code{plotAvgMS1}.
}

