% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpec_modified.R
\name{getSpec_modified}
\alias{getSpec_modified}
\title{Modification to xcms::getSpec()}
\usage{
getSpec_modified(object, scanrange = NULL, ...)
}
\arguments{
\item{prmz}{A numeric value. Numeric vectors cannot be used
(in that cases, only the first element will be used).}

\item{xraw}{An object of class \code{xcmsRaw}, which is read into by \code{xcmsRaw()}
with \code{includeMSn = TRUE} option.}

\item{mzabs}{Search tolerance in m/z (not ppm).}
}
\value{
A data frame with columns of the names \code{mz} and \code{intensity}.
}
\description{
A function that extract scan(s) from a \code{xcmsRaw} object.
   A bug that prevented scan averaging of MS2 data was removed.
   Incontrast to the original getSpec, scanrange must be provided (default NULL).
}

