% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.avgMS2.R
\name{plot.avgMS2}
\alias{plot.avgMS2}
\title{Plot averaged MS2 spectrum for a selected m/z or scans}
\usage{
\method{plot}{avgMS2}(xcmsraw, mz = NULL, scanNum = NULL, mzabs = 0.1,
  digits = 4, mzrange = NULL, halfWindowSize = NULL, SNR = 5,
  top = 10, noisePeaks = NULL, ...)
}
\arguments{
\item{xcmsraw}{xcmsRaw object}

\item{mz}{Precursor m/z value to plot (default NULL). Either one of the  \code{mz}
or \code{scanNum} is used.}

\item{scanNum}{Scan numbers to plot (default NULL). A numeric vector.}

\item{mzabs}{Tolerance for scan selection in absolute m/z values.}

\item{digits}{Digits for peak label.}

\item{mzrange}{m/z range to plot.}

\item{halfWindowSize}{See \code{MALDIquant::labelPeaks()} for detail.}

\item{SNR}{See \code{MALDIquant::labelPeaks()} for detail.}

\item{top}{The number of peaks to label (default 10).}

\item{noisePeaks}{A numeric vector. Peaks that should not be labeled.}
}
\value{
Invisibly (only when assigned to a variable) returns averaged spectrum
     as a data frame. If no MS/MS spectrum is found for the precursor m/z or
     scan numbers, NULL is returned.
}
\description{
From xcmsRaw object, this function allows plotting MS2 spectrum for a selected m/z
   or selected scans, with peak labels and PrecursorMz line. Additionally, common noise
   peaks can be excluded from the peak labels.
}

