% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIsotope.R
\name{plotIsotope}
\alias{plotIsotope}
\title{Plot isotope patterns}
\usage{
plotIsotope(chemsc, z = 0, add = FALSE, type = "p", ymax = 1,
  max.iso = 5, ...)
}
\arguments{
\item{chemsc}{Chemical sum composition (elemental formula) as a character (e.g. "H2O").}

\item{z}{Charge}

\item{add}{Logical (default FALSE). If TRUE, overlay isotope pattern on an existing plot.}

\item{type}{Type argument for plotting (used within points() function). Default, "p" (point).}

\item{ymax}{Default=1. The height of the maximum peak.}

\item{max.iso}{Maximum number of isotopes to plot. Default, 5.}
}
\description{
A wrapper function for Rdisop::getMolecule, getIsotopes etc., that allows easier plotting of the isotope patterns.
}

