% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPIS.R
\name{plotPIS}
\alias{plotPIS}
\title{Plot PIS data set (comparison of two spectra)}
\usage{
plotPIS(pisdata, col = c("blue", "red"), lty = c(1, 1), mzrange = NULL,
  fill.col = adjustcolor(c("skyblue", "salmon"), alpha.f = 0.5), ...)
}
\arguments{
\item{pisdata}{A data frame constructed from xcmsRaw object via \code{constructPIS()} function.}

\item{col}{A vector of colors for lines. Default, blue and red.}

\item{lty}{Types of lines. Same as base graphics.}

\item{mzrange}{The m/z range to plot. A numeric vector of the length 2.}

\item{fill.col}{A vector of colors for area under the curve. Default, transparent colors.}
}
\description{
Plot PIS (precursor ion scanning) data
}

