% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSliceMS2.R
\name{plotSliceMS2}
\alias{plotSliceMS2}
\title{Plot "precursor ion scan"- or "neutral loss scan"-like spectrum from DDA xcmsRaw object.}
\usage{
plotSliceMS2(xraw, mz, ppm = 50, mzrange = NULL, ylim = NULL,
  slicemode = "FRAG", scanType = "profile", plotSpec = TRUE,
  type = c("h", "p"), col = c("blue", "red"), pch = 1, cex = 0.5,
  labelPeaks = TRUE, top = 5, SNR = 5, halfWindowSize = NULL,
  digits = 4, ...)
}
\arguments{
\item{xraw}{\code{xcmsRaw} object.}

\item{mz}{\code{xcmsRaw} object.}

\item{ppm}{Numeric (default 10 ppm). Integration range. If \code{mzabs} is provided,
this parameter will not be ignored.}

\item{slicemode}{Character (default "FRAG").Select one from "FRAG" or "NL",
to plot precursor ion spectrum with fragment or neutral loss intensity respectively.}

\item{scanType}{Character (default "profile"). Either "profile" or "centroid".}

\item{plotSpec}{Logical (default TRUE). Whether plot pseudo-precursor ion spectrum.
If set to FALSE, only returns pseudo precursor ion spectrum as a matrix.}

\item{type}{A character vector containing \code{type} argments that are used in standard
\code{plot()}. For example, \code{c("h", "p")} results in points over vertical lines.}

\item{col}{Colors used to plot multiple spectra. A character vector (default \code{c("blue", "red")})}

\item{pch}{Same as \code{plot()}. Type of points.}

\item{cex}{Same as \code{plot()}. Size of points.}

\item{labelPeaks}{Logical (default TRUE), that determines whether to plot peak labels.}

\item{top}{The number of peak labels (default 5).
By default, top 5 peaks in each spectrum will be labelled.}

\item{SNR}{S/N ratio for peak picking (default 5).}

\item{halfWindowSize}{A parameter for peak picking (default NULL).
By default, automatically determined based on data.
Larger value prevents peak labeling of nearer peaks.}

\item{digits}{Digits of peak label (default 4).}
}
\value{
Plot spectrrum. Plus, if assigned to a variable, returns a data.frame
     with precursor m/z, and intensities (AUC for profile, sum for cetroid).
}
\description{
Note that multiple scans with the same precursor m/z values are averaged
     (after peak integration).
}

