% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMgf.R
\name{writeMgf}
\alias{writeMgf}
\title{A function that export mgf file from filename, precursor mz, and a peak table.}
\usage{
writeMgf(file, precursor, peak.table, charge = "+1")
}
\arguments{
\item{file}{A file name for MGF file to be exported.}

\item{precursor}{A precursor m/z value.}

\item{peak.table}{A data.frame with 2 column (mz, int).}

\item{charge}{Charge of the precursor. Character (default "+1").}
}
\description{
A function that export mgf file from filename, precursor mz, and a peak table.
}

