% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsAvgHyperSpec.R
\name{xcmsAvgHyperSpec}
\alias{xcmsAvgHyperSpec}
\title{A function that retrieve averaged spectra from xcmsRaw object as a hyperSpec object}
\usage{
xcmsAvgHyperSpec(xcmsraw, scans, step = 0.01, mzrange = NULL)
}
\arguments{
\item{xcmsraw}{An object of class \code{xcmsRaw} class}

\item{scans}{A numeric vector}

\item{step}{Numeric (default 0.01). The step size of x axis to which
the raw data is interpolated.}

\item{mzrange}{Range of m/z to be included in the output \code{hyperSpec} object.
By default, the min and max values of m/z in the data are used.}
}
\value{
Returns an object of class \code{hyperSpec}.
}
\description{
A function that retrieve averaged spectra from xcmsRaw object as a hyperSpec object
}

