% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsrawAuc.R
\name{xcmsrawAuc}
\alias{xcmsrawAuc}
\title{Calculate area under the curve from \code{xcmsRaw} object}
\usage{
xcmsrawAuc(scan, mzrange = NULL, mz = NULL, mzabs = NULL, xraw,
  type = "profile", msLevel = 1)
}
\arguments{
\item{scan}{An Numeric vector. Scan index within the indicated MS level..
Note that this scan index is NOT the actual scan number (include both MS1 and MS2).
This corresoinds to the index of the actual scan number in
\code{xcmsRaw@msnAcquisitionNum} slot.}

\item{mz}{Center m/z values}

\item{mzabs}{Width to integrate}

\item{xraw}{An object of class \code{xcmsRaw} class}

\item{type}{"profile" or "centroid". Default, "profile".}

\item{msLevel}{MS level (default 1).}
}
\value{

}
\description{
Calculate area under the curve from \code{xcmsRaw} object
}

