% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsTICHyperSpec.R
\name{xcmsTICHyperSpec}
\alias{xcmsTICHyperSpec}
\title{A function that retrieve TIC from a list of xcmsRaw objects as a hyperSpec object}
\usage{
xcmsTICHyperSpec(xraw_list, rtstep = 1, rtrange = NULL, mzrange = NULL,
  mzstep = 1)
}
\arguments{
\item{xraw_list}{A list of objects of class \code{xcmsRaw} class}

\item{rtrange}{Range of m/z to be included in the output \code{hyperSpec} object.
By default, the min and max values of m/z in the data are used.}

\item{step}{Numeric (default 1 sec). The step size of retention time axis to which
the raw data is interpolated.}
}
\value{
Returns an object of class \code{hyperSpec}.
}
\description{
From a list of xcmsRaw objects (LC-MS run), this function returns a hyperSpec object (TIC).
}

