% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsAvgHyperSpec.R
\name{xcmsAvgHyperSpec}
\alias{xcmsAvgHyperSpec}
\title{A function that retrieve averaged spectra from a xcmsRaw object as a hyperSpec object}
\usage{
xcmsAvgHyperSpec(xcmsraw, scans = NULL, msLevel = 1, step = 0.01,
  mzrange = NULL)
}
\arguments{
\item{xcmsraw}{An object of class \code{xcmsRaw} class}

\item{scans}{A numeric vector (default NULL returns average of all the scans). Scan index within the ms level, not an actual scan number.}

\item{msLevel}{MS level (MS1 or MS2). Either 1 or 2. Default, 1 (MS1).}

\item{step}{Numeric (default 0.01). The step size of x axis to which
the raw data is interpolated.}

\item{mzrange}{Range of m/z to be included in the output \code{hyperSpec} object.
By default, the min and max values of m/z in the data are used.}
}
\value{
Returns an object of class \code{hyperSpec}.
}
\description{
From a xcmsRaw object (with multiple scans), this function returns a hyperSpec object (averaged spectrum). As this function only accepts one object, later merging would be necessary to construct a hyperSpec object with multiple spectra (use rbind()).
}

